application/pyproject.toml
[project]
name = "led-matrix-controller"
version = "1.0.0"
description = "Control 128x32 HUB75 LED matrix panel via USB CDC"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.20.0",
    "pyserial>=3.5",
    "opencv-python>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
]

[project.scripts]
led-matrix = "led_matrix_controller.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/led_matrix_controller"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
